% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractsofgroup.R
\name{abstractsofgroup}
\alias{abstractsofgroup}
\title{A Function that collects the PMID's of a specific clustering group, after clustering with screenmed function.}
\usage{
abstractsofgroup(clustering, group_number)
}
\arguments{
\item{clustering}{The $clustering output of the screemed function.}

\item{group_number}{The number of the group the screenmed divided the initial_search.}
}
\description{
Gathering the PMIDs of abstracts which belong to a specific cluster.
}
