% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_clean.R
\name{diagnosis_clean}
\alias{diagnosis_clean}
\title{Clean and manipulate the abstracts}
\usage{
diagnosis_clean(x)
}
\arguments{
\item{x}{}
}
\description{
By Using tm functions works for cleaning, stemming removing
special words etc.
}
\examples{

}
