% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_by_name.R
\name{mesh_by_name}
\alias{mesh_by_name}
\title{A function that provides the PMID numbers of publications in terms of the names of their Descriptors & Qualifiers}
\usage{
mesh_by_name(x, Descriptor, Qualifier)
}
\arguments{
\item{x}{A character vector containing the PMID numbers of the articles of our search}

\item{Descriptor}{A character vector containing the mesh Descriptors.}

\item{Qualifier}{A character vector containing the mesh Qualifiers.}
}
\description{
It can provide all the PMIDS of publications with specific Descriptors
and Qualifiers. It has a 300 Abstracts input limit (x < 300)
}
