% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_by_name_bq.R
\name{mesh_by_name_bq}
\alias{mesh_by_name_bq}
\title{A function that provides the PMID numbers of publications in terms of the names of their Descriptors & Qualifiers}
\usage{
mesh_by_name_bq(x, d, q)
}
\arguments{
\item{x}{A character vector containing the PMID numbers of the articles of our search}

\item{Descriptor}{A character vector containing the mesh Descriptors.}

\item{Qualifier}{A character vector containing the mesh Qualifiers.}
}
\description{
It can provide all the PMIDS of publications with specific Descriptors
and Qualifiers.
}
\examples{

initial_search <- read.csv(system.file("extdata", "csv-randomized-set.csv", package = "screenmedR"))
initialPMID<-initial_search$PMID
Descriptor<-c("Blood Pressure","Dobutamine","Humans","Infant, Newborn")
Qualifier<-c("administration & dosage")
mesh_by_name_bq(initialPMID,Descriptor,Qualifier)

}
