% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_clean.R
\name{mesh_clean}
\alias{mesh_clean}
\title{A function that screens relevant publications in terms of common numbers of mesh terms.}
\usage{
mesh_clean(x, y, d, q)
}
\arguments{
\item{x}{A character vector containing the PMID numbers of the articles of our search}

\item{y}{A character vector containing the pmid numbers of publications that the user would like to compare with x.}

\item{d}{An integer that describes the number of Descriptors that the user wants the two vectors to have at least in common.}

\item{q}{An integer that describes the number of Qualifiers that the user wants the two vectors to have at least in common.}
}
\value{
result_final
}
\description{
It provides the PMID numbers of a set of publications x which have at least d descriptors and q qualifiers in common with a small default set y. It has a limit of 300 Abstracts (x< 300).
}
