% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_clean_bq.R
\name{mesh_clean_bq}
\alias{mesh_clean_bq}
\title{Screening-cleaning relevant publications in terms of number of mesh terms.}
\usage{
mesh_clean_bq(x, y, d, q)
}
\arguments{
\item{x}{A character vector containing the PMID numbers of the articles of our search}

\item{y}{A character vector containing the pmid numbers of publications that the user would like to compare with x.}

\item{d}{An integer that describes the number of Descriptors that the user wants the two vectors to have at least in common.}

\item{q}{An integer that describes the number of Qualifiers that the user wants the two vectors to have at least in common.}
}
\description{
It provides the PMID numbers of a set of publications x which have at least d descriptors and q qualifiers in common with a small default set y.
}
\examples{

initial_search <- read.csv(system.file("extdata", "csv-randomized-set.csv", package = "screenmedR"))
initialPMID<-initial_search$PMID
knownPMID<-c("18822428","8276025","16452355","17329276","8346957")
mesh_clean_bq(initialPMID,knownPMID,11,2)

}
