% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractsofgroup.R
\name{abstractsofgroup}
\alias{abstractsofgroup}
\title{Screens Pubmed PMID numbers in terms of those of 5 initial ones}
\usage{
abstractsofgroup(clustering, group_number, initial_search)
}
\arguments{
\item{clustering, group_number, initial_search}{}
}
\description{
This package predicts which abstracts are closer in terms of words with an initial group of n abstracts that the user provides.
}
\examples{

}
