% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractsofgroup.R
\name{abstractsofgroup}
\alias{abstractsofgroup}
\title{Abstract PMID collector of a specific group, after clustering.}
\usage{
abstractsofgroup(clustering, group_number, initial_search)
}
\arguments{
\item{clustering, group_number, initial_search}{}
}
\description{
Gathering the PMIDs of abstracts belonging to a specific cluster.
}
\examples{

}
