% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_clean.R
\name{diagnosis_clean}
\alias{diagnosis_clean}
\title{Screens Pubmed PMID numbers in terms of those of 5 initial ones}
\usage{
diagnosis_clean(x)
}
\arguments{
\item{x}{}
}
\description{
This package predicts which abstracts are closer in terms of words with an initial group of n abstracts that the user provides.
}
\examples{

}
