% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_by_name.R
\name{mesh_by_name}
\alias{mesh_by_name}
\title{Screens Pubmed PMID numbers in terms of the names of Descriptors & Qualifiers}
\usage{
mesh_by_name(x, Descriptor, Qualifier)
}
\arguments{
\item{x, Descriptor, Qualifier}{}
}
\description{
It can provide all the PMIDS of publications with specific Descriptors
and Qualifiers. It is used for x < 300 PMIDS
}
\examples{

}
