% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_by_name_bq.R
\name{mesh_by_name_bq}
\alias{mesh_by_name_bq}
\title{Screens Pubmed PMID numbers in terms of the names of Descriptors & Qualifiers}
\usage{
mesh_by_name_bq(x, d, q)
}
\arguments{
\item{x, Descriptor, Qualifier}{}
}
\description{

}
\examples{

}
