% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_clean.R
\name{mesh_clean}
\alias{mesh_clean}
\title{Cleaning not relevant publications in terms of number of mesh terms.}
\usage{
mesh_clean(x, y, d, q)
}
\arguments{
\item{x, y, d, q}{}
}
\value{
result_final
}
\description{
It provides the PMID numbers of a set of publications x which have at least
 d descriptors and q qualifiers in common with a small default set y. Should be used for
 x < 300 abstracts.
}
\examples{

}
