% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_clean.R
\name{mesh_clean}
\alias{mesh_clean}
\title{Screens Pubmed PMID numbers in terms of those of 5 initial ones}
\usage{
mesh_clean(x, y, d, q)
}
\arguments{
\item{x, y, d, q}{}
}
\description{
This package predicts which abstracts are closer in terms of words with an initial group of n abstracts that the user provides.
}
\examples{

}
