% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmed.R
\name{screenmed}
\alias{screenmed}
\title{Screens Pubmed PMID numbers in accordance of a small group of relevant publications}
\usage{
screenmed(initial_search, filtered, sparcity, group_number)
}
\arguments{
\item{initial_search, filtered, sparcity, group_number}{}
}
\value{
group_number,cosine_similarity,clustering,missing_abstracts
}
\description{
It provides us results for the cosine similarity of the group_number clusters
as well as a list of the clustered abstracts and in which cluster they belong. Finally it
also provides the PMIDS of the publications that no abstract was found.
}
\examples{

}
