% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenmed.R
\name{screenmed}
\alias{screenmed}
\title{Screens Pubmed PMID numbers in terms of those of 5 initial ones}
\usage{
screenmed(initial_search, filtered, sparcity, group_number)
}
\arguments{
\item{initial_search, filtered, sparcity, group_number}{}
}
\description{
This package predicts which abstracts are closer in terms of words with an initial group of n abstracts that the user provides.
}
\examples{

}
